+++ app/src/main/java/com/rezolve/gistscanner/ui/landing/LandingActivity.java
@@ -1,9 +1,15 @@
 package com.rezolve.gistscanner.ui.landing;
 
+import android.Manifest;
 import android.content.Intent;
+import android.content.pm.PackageManager;
+import android.support.annotation.NonNull;
+import android.support.v4.app.ActivityCompat;
+import android.support.v4.content.ContextCompat;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
+import android.widget.Toast;
 
 import com.rezolve.gistscanner.R;
 import com.rezolve.gistscanner.ui.main.MainActivity;
@@ -11,6 +17,7 @@ import com.rezolve.gistscanner.ui.main.MainActivity;
 import com.rezolve.gistscanner.ui.util.UIUtils;
 
 public class LandingActivity extends AppCompatActivity {
+    private static final int PERMISSION_REQUEST_CODE = 1234;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -21,7 +28,33 @@ public class LandingActivity extends AppCompatActivity {
     }
 
     public void launchMainActivity(View view) {
-        startActivity(new Intent(this, MainActivity.class));
-        finish();
+        checkForPermissions();
+    }
+
+    private void checkForPermissions() {
+        if (ContextCompat.checkSelfPermission(this, Manifest.permission.CAMERA)
+                != PackageManager.PERMISSION_GRANTED) {
+            ActivityCompat.requestPermissions(
+                    this,
+                    new String[]{Manifest.permission.CAMERA}, PERMISSION_REQUEST_CODE);
+        } else {
+            startActivity(new Intent(this, MainActivity.class));
+            finish();
+        }
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        if (PERMISSION_REQUEST_CODE == requestCode) {
+            if (grantResults.length > 0
+                    && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                startActivity(new Intent(this, MainActivity.class));
+                finish();
+            } else {
+                Toast.makeText(this, R.string.camera_permission_required,
+                        Toast.LENGTH_SHORT).show();
+            }
+        }
     }
 }
