+++ app/src/main/java/com/rezolve/gistscanner/ui/main/MainActivity.java
@@ -1,11 +1,7 @@
 package com.rezolve.gistscanner.ui.main;
 
-import android.Manifest;
-import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
 import android.text.TextUtils;
 import android.widget.Toast;
 
@@ -23,8 +19,6 @@ import com.rezolve.gistscanner.ui.util.UIUtils;
 public class MainActivity extends DaggerAppCompatActivity implements
         ScannerFragment.OnFragmentInteractionListener {
 
-    private static final int PERMISSION_REQUEST_CODE = 1234;
-
     @Inject
     MainFragment mainFragment;
 
@@ -40,20 +34,13 @@ public class MainActivity extends DaggerAppCompatActivity implements
             getSupportActionBar().setDisplayHomeAsUpEnabled(true);
         }
 
-        if (ContextCompat.checkSelfPermission(this, Manifest.permission.CAMERA)
-                != PackageManager.PERMISSION_GRANTED) {
-            ActivityCompat.requestPermissions(
-                    this,
-                    new String[]{Manifest.permission.CAMERA}, PERMISSION_REQUEST_CODE);
-        } else {
-            if (savedInstanceState == null) {
-                ScannerFragment fragment = ScannerFragment.class.cast(
-                        getSupportFragmentManager().findFragmentById(R.id.container)
-                );
+        if (savedInstanceState == null) {
+            ScannerFragment fragment = ScannerFragment.class.cast(
+                    getSupportFragmentManager().findFragmentById(R.id.container)
+            );
 
-                if (fragment == null) {
-                    addScannerFragment();
-                }
+            if (fragment == null) {
+                addScannerFragment();
             }
         }
     }
@@ -75,20 +62,6 @@ public class MainActivity extends DaggerAppCompatActivity implements
                         mainFragment, R.id.container, MAIN_FRAGMENT_TAG);
             }
         });
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        if (PERMISSION_REQUEST_CODE == requestCode) {
-            if (grantResults.length > 0
-                    && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                addScannerFragment();
-            } else {
-                Toast.makeText(this, R.string.camera_permission_required,
-                        Toast.LENGTH_SHORT).show();
-            }
-        }
     }
 
     @Override
